README: Add note for copying gtk icon
By default, the windows system does not have the icons required by gtk,
and it needs to be manually copied to the corresponding directory.
Otherwise, many icons (such as the maximized window icon) will be blank
when running the gtk program, and the following warning will appear:

Gtk-WARNING **: Could not find the icon 'window-minimize-symbolic-ltr'. The 'hicolor' theme
was not found either, perhaps you need to install it.

Signed-off-by: Dunrong Huang <riegamaths@gmail.com>